# openapi.yaml
openapi: 3.1.0
info:
  title: Audicle
  version: 0.57.0
paths:
...
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Post Login
      operationId: post_login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthActionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/auth/password:
    put:
      tags:
      - auth
      summary: Put Password
      operationId: put_password_api_v1_auth_password_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthActionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/auth/revoke-sessions:
    post:
      tags:
      - auth
      summary: Revoke Sessions
      operationId: revoke_sessions_api_v1_auth_revoke_sessions_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: boolean
                type: object
                title: Response Revoke Sessions Api V1 Auth Revoke Sessions Post
      security:
      - {}
      - AdminSession: []
        CsrfToken: []
...
  /api/v1/episodes:
    get:
      tags:
      - episodes
      summary: List Episodes
      operationId: list_episodes_api_v1_episodes_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 25
          title: Per Page
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Q
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EpisodeListItem'
                title: Response List Episodes Api V1 Episodes Get
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - AdminSession: []
  /api/v1/episodes/{episode_id}/chapters:
    post:
      tags:
      - episodes
      summary: Regenerate Chapters
      description: 'Rebuild chapters for a finished episode from its stored transcript.


        Cues are one per TTS chunk, so the transcript already carries the timeline

        the chapter stage used; no TTS or re-encode is needed. Existing chapters

        survive a failed run, and the episode GUID is left alone because the audio

        does not change.'
      operationId: regenerate_chapters_api_v1_episodes__episode_id__chapters_post
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          title: Episode Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChaptersRegenResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - AdminSession: []
        CsrfToken: []
  /api/v1/episodes/{episode_id}:
    delete:
      tags:
      - episodes
      summary: Delete Episode
      operationId: delete_episode_api_v1_episodes__episode_id__delete
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          title: Episode Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteEpisodeResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - AdminSession: []
        CsrfToken: []
...
components:
  schemas:
    AuthActionResponse:
      properties:
        authenticated:
          type: boolean
          title: Authenticated
        password_set:
          type: boolean
          title: Password Set
        csrf_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Csrf Token
      additionalProperties: false
      type: object
      required:
      - authenticated
      - password_set
      title: AuthActionResponse
...
    ChaptersRegenResponse:
      properties:
        episode_id:
          type: string
          title: Episode Id
        chapter_count:
          type: integer
          title: Chapter Count
      additionalProperties: false
      type: object
      required:
      - episode_id
      - chapter_count
      title: ChaptersRegenResponse
...
    DeleteEpisodeResponse:
      properties:
        id:
          type: string
          title: Id
        files_removed:
          type: integer
          minimum: 0.0
          title: Files Removed
      additionalProperties: false
      type: object
      required:
      - id
      - files_removed
      title: DeleteEpisodeResponse
    EpisodeListItem:
      properties:
        id:
          type: string
          title: Id
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        original_url:
          type: string
          title: Original Url
        audio_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Audio Path
        audio_size_bytes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Audio Size Bytes
        artwork_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Artwork Path
        duration_secs:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Secs
        pub_date:
          type: string
          title: Pub Date
        updated_at:
          type: string
          title: Updated At
        has_cleaned_text:
          type: boolean
          title: Has Cleaned Text
        has_chapters:
          type: boolean
          title: Has Chapters
        source_type:
          type: string
          title: Source Type
        source_filename:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Filename
        voice_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Voice Label
      additionalProperties: false
      type: object
      required:
      - id
      - title
      - author
      - original_url
      - audio_path
      - audio_size_bytes
      - artwork_path
      - duration_secs
      - pub_date
      - updated_at
      - has_cleaned_text
      - has_chapters
      - source_type
      - source_filename
      - voice_label
      title: EpisodeListItem
...
    LoginRequest:
      properties:
        password:
          type: string
          maxLength: 200
          minLength: 1
          title: Password
      additionalProperties: false
      type: object
      required:
      - password
      title: LoginRequest
...
    PasswordRequest:
      properties:
        current_password:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Current Password
        new_password:
          type: string
          maxLength: 200
          title: New Password
      additionalProperties: false
      type: object
      required:
      - new_password
      title: PasswordRequest
...
    ErrorEnvelope:
      type: object
      required:
      - error
      - status
      properties:
        error:
          type: string
        status:
          type: integer
        details:
          type: object
          additionalProperties: true
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: audicle_session
      description: Required for the admin API when a password is configured.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with the admin session for unsafe methods.
    FeedKey:
      type: apiKey
      in: query
      name: key
      description: Required for protected feeds and non-artwork media.